move LargeModal to generic Modal file
