+++ src/components/game/LogModal/LogModal.tsx
@@ -3,7 +3,8 @@ import type { BasicPlayer } from "types/game"
 import type { GameLog } from "./log.controller"
 import LogToolbar from "./LogToolbar/LogToolbar"
 import LogEntry from "./LogEntry"
-import { LargeModal, LogContainer, ErrorContainer } from "./LogModalStyles"
+import { LogContainer, ErrorContainer } from "./LogModalStyles"
+import { LargeModal } from "components/base/common/Modal"
 
 type Props = {
   players: BasicPlayer[],
